scrapy_app/rabbitmq_broker/requirements.txt
amqp==2.6.0
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.7
async-exit-stack==1.0.1
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
billiard==3.6.3.0
celery==4.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
Django==3.0.5
django-cors-headers==3.4.0
django-rest-framework-mongoengine==3.4.1
djongo==1.3.3
ecdsa==0.15
elasticsearch==7.8.1
fastapi==0.55.1
filelock==3.0.12
h11==0.9.0
httptools==0.1.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
items==0.6.5
kombu==4.6.10
lxml==4.5.1
mongoengine==0.20.0
nulltype==2.3.1
parsel==1.6.0
passlib==1.7.2
pika==1.1.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.5.1
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pymongo==3.10.1
pyOpenSSL==19.1.0
python-dotenv==0.13.0
python-jose==3.1.0
python-multipart==0.0.5
python-scrapyd-api==2.1.2
pytz==2020.1
queuelib==1.5.0
requests==2.23.0
rsa==4.0
Scrapy==2.1.0
scrapyd==1.2.1
scrapyd-client==1.1.0
service-identity==18.1.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.2.4
starlette==0.13.2
Twisted==20.3.0
tzlocal==2.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
vine==1.3.0
virtualenv==20.0.21
w3lib==1.22.0
websockets==8.1
zipp==3.1.0
zope.interface==5.1.0
